==> notebooks/01_exploration_spark.ipynb <==
# Etape 1 - Exploration et chargement Spark

**Objectif** : Charger et explorer les donnees de qualite de l'air avec PySpark

---
---

## Imports

import os
from pathlib import Path
from pyspark.sql import SparkSession
from pyspark.sql import functions as F

---

## Chemins des données

DATA_DIR = (Path.cwd() / ".." / "data").resolve()
AIR_QUALITY_PATH = os.path.join(DATA_DIR, "air_quality_raw.csv")
STATIONS_PATH = os.path.join(DATA_DIR, "stations.csv")
WEATHER_PATH = os.path.join(DATA_DIR, "weather_raw.csv")

---

## 1.1 Création d'une session Spark locale

## Creer une session Spark locale
spark = SparkSession.builder \
    .appName("TP Qualite Air - Exploration") \
    .master("local[*]") \
    .config("spark.driver.memory", "4g") \
    .config("spark.sql.shuffle.partitions", "8") \
    .getOrCreate()

## Reduire les logs
spark.sparkContext.setLogLevel("WARN")

## Affichage de l'actuelle version de Spark & de l'url de Spark UI
print(f"Spark version: {spark.version}")
print(f"Spark UI: {spark.sparkContext.uiWebUrl}")

---

## 1.2 Chargement du fichier air_quality_raw.csv en DataFrame Spark (les données de la qualité de l'air)

## Chargement du CSV avec inference de schema
df_air_raw = spark.read \
    .option("header", "true") \
    .option("inferSchema", "true") \
    .csv(AIR_QUALITY_PATH)

print("- Fichier de la qualité de l'air 'air_quality_raw.csv' a été chargé avec succès.")

## Information sur le nombre de lignes et de colonnes
print("- Nombre de lignes :", df_air_raw.count())
print("- Nombre de colonnes :", len(df_air_raw.columns))

## Apercu des donnees
print("- Apercu des donnees :")
df_air_raw.show(10, truncate=False)

---

## 1.3 Affichage du schéma inféré et identification des problèmes de typage

### 1.3.1 Affichage du schéma inféré

df_air_raw.printSchema()

### 1.3.2 Identification des problèmes de typage
- **timestamp**

la colonne ``timestamp`` est en string car Il y'a differents formats de timestamp :

print("Exemples de formats de timestamp:")
df_air_raw.select("timestamp").distinct().show(20, truncate=False)

- **value**

La colonne ``value`` est en string également car elle contient des valeurs avec des virgules et des valeurs textuelles :

## Les valeurs qui ne peuvent pas etre converties en nombre
df_non_numeric = df_air_raw.filter(
    ~F.col("value").rlike("^-?[0-9]+[.,]?[0-9]*$")
)

print(f"- Nombre de valeurs non numeriques : {df_non_numeric.count():,}")

print(f"- Differents valeurs non numeriques :")
df_non_numeric.select("value").distinct().show()

## Valeurs avec virgule comme separateur decimal
df_with_comma = df_air_raw.filter(F.col("value").contains(","))
print(f"- Nombre de valeurs avec virgule : {df_with_comma.count():,}")

## Affichages d'exemples
print(f"- Exemples :")
df_with_comma.select("value").show(5)

---

## 1.4 Calcul des statistiques descriptives par polluant

Initialement on a la colonne ``value`` en string. 

Pour calculer les statistiques descriptives, on va créer une dataFrame 'df_air_numeric' qui contient les mêmes données que 'df_air_raw' plus une colonne 'value_clean' qui contient les valeurs de la colonne 'value_clean' converties en double.

## Nombre des valeurs Nulls initiales dans df_air_raw
df_air_raw.filter(F.col("value").isNull()).count()

## Ajouter de la colonne 'value_clean' dans la dataFrame df_air_raw
## & convertion de value en double (en remplacant la virgule par un point)
df_air_numeric = df_air_raw.withColumn(
    "value_clean",
    F.regexp_replace(F.col("value"), ",", ".").cast("double")
)

## Apercu de la df_air_numeric
print("- Apercu de la df_air_numeric :")
df_air_numeric.show(10)

## Nombre des valeurs Nulls dans df_air_numeric
print("- Nombre des valeurs Nulls dans la colonne 'value_clean' :", df_air_numeric.filter(F.col("value_clean").isNull()).count())

## Calcul des statistiques descriptives par polluant (en ignorant les valeurs nulles de 'value_clean')
stats_by_pollutant = df_air_numeric.filter(F.col("value_clean").isNotNull()) \
    .groupBy("pollutant") \
    .agg(
        F.count("*").alias("count"),
        F.round(F.mean("value_clean"), 2).alias("mean"),
        F.round(F.stddev("value_clean"), 2).alias("stddev"),
        F.round(F.min("value_clean"), 2).alias("min"),
        F.round(F.max("value_clean"), 2).alias("max"),
        F.round(F.expr("percentile(value_clean, 0.5)"), 2).alias("median")
    ) \
    .orderBy("pollutant")

print("\n- Statistiques par polluant (en ignorant les valeurs nulles) :")
stats_by_pollutant.show()

## Identification des valeurs aberrantes
print("- Valeurs negatives:")
df_air_numeric.filter(F.col("value_clean") < 0).groupBy("pollutant").count().show()

print("- Valeurs > 1000 ug/m3:")
df_air_numeric.filter(F.col("value_clean") > 1000).groupBy("pollutant").count().show()

---

## 1.5 Comptage des valeurs nulles par colonne

null_counts = df_air_raw.select([
    F.count(F.when(F.col(c).isNull() | (F.col(c) == ""), c)).alias(c)
    for c in df_air_raw.columns
])

print("- Nombre de valeurs nulles/vides par colonne:")
null_counts.show()

---

## 1.6 Identification des stations avec le plus d'enregistrements

## station ids avec le plus d'enregistrements (affichage par ordre décroissant du nombre des enregistrements)
print("- Les stations ids avec le plus d'enregistrements :")
df_stations_enregistrements_count = df_air_raw.groupBy('station_id').count().withColumnRenamed("count", "nbre_d_enregistrements").orderBy(F.col("count").desc())
df_stations_enregistrements_count.show(10)

## Chargement du CSV des stations avec inference de schema
df_stations = spark.read \
    .option("header", "true") \
    .option("inferSchema", "true") \
    .csv(STATIONS_PATH)

print("- Fichier de des stations 'stations.csv' a été chargé avec succès.")

## Information sur le nombre de lignes et de colonnes
print("- Nombre de lignes :", df_stations.count())
print("- Nombre de colonnes :", len(df_stations.columns))

## Apercu des donnees
print("- Apercu des donnees :")
df_stations.show(10, truncate=False)

## Affichage des stations (avec leurs données) avec le plus d'enregistrements (ordre décroissant)
df_stations_enregistrements_count.join(df_stations, on="station_id", how="left").orderBy(F.col("nbre_d_enregistrements").desc()).show()

---

## 1.7 Synthese des problemes de qualite identifies

## Nombre total des enregistrements
total_enregistrements = df_air_raw.count()

## Valeurs non numeriques
non_numeric = df_air_raw.filter(
    ~F.col("value").rlike("^-?[0-9]+[.,]?[0-9]*$")
).count()

## Valeurs avec virgule
with_comma = df_air_raw.filter(F.col("value").contains(",")).count()

## Valeurs negatives (apres conversion)
negative = df_air_numeric.filter(F.col("value_clean") < 0).count()

## Valeurs aberrantes > 1000
outliers = df_air_numeric.filter(F.col("value_clean") > 1000).count()

## Doublons
duplicates = total_enregistrements - df_air_raw.dropDuplicates(["station_id", "timestamp", "pollutant"]).count()


print(f"Total enregistrements : {total_enregistrements:,}")
print()
print(f"Problemes identifies :")
print(f"  - Valeurs non numeriques: {non_numeric:,} ({non_numeric/total_enregistrements*100:.2f}%)")
print(f"  - Valeurs avec virgule decimale: {with_comma:,} ({with_comma/total_enregistrements*100:.2f}%)")
print(f"  - Valeurs negatives: {negative:,} ({negative/total_enregistrements*100:.2f}%)")
print(f"  - Valeurs aberrantes (>1000): {outliers:,} ({outliers/total_enregistrements*100:.2f}%)")
print(f"  - Doublons: {duplicates:,} ({duplicates/total_enregistrements*100:.2f}%)")
print(f"  - Formats de dates multiples : 4 formats differents detectes")

---

## Fermer la session Spark

spark.stop()
